# Assignment8/src/api.yaml
openapi: 3.0.3
info:
  title: E-commerce API
  description: API for managing products, shopping carts, warehouse operations, and credit card processing
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
paths:
...
  /products/{productId}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve a product's details using its unique identifier
      operationId: getProduct
      parameters:
        - name: productId
          in: path
          required: true
          description: Unique identifier for the product
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        '200':
          description: Product found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /products/{productId}/details:
    post:
      tags:
        - Products
      summary: Add product details
      description: Add or update detailed information for a specific product
      operationId: addProductDetails
      parameters:
        - name: productId
          in: path
          required: true
          description: Unique identifier for the product
          schema:
            type: integer
            format: int32
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '204':
          description: Product details added successfully
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Shopping Cart Service Endpoints
...
components:
  schemas:
    Product:
      type: object
      required:
        - product_id
        - sku
        - manufacturer
        - category_id
        - weight
      properties:
        product_id:
          type: integer
          format: int32
          minimum: 1
          description: Unique identifier for the product
          example: 12345
        sku:
          type: string
          minLength: 1
          maxLength: 100
          description: Stock Keeping Unit - unique product code
          example: "ABC-123-XYZ"
        manufacturer:
          type: string
          minLength: 1
          maxLength: 200
          description: Product manufacturer name
          example: "Acme Corporation"
        category_id:
          type: integer
          format: int32
          minimum: 1
          description: Product category identifier
          example: 456
        weight:
          type: integer
          format: int32
          minimum: 0
          description: Product weight in grams
          example: 1250
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "INVALID_INPUT"
        message:
          type: string
          description: Human-readable error message
          example: "The provided input data is invalid"
        details:
          type: string
          description: Additional error details
          example: "Product ID must be a positive integer"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...